+++ templates/react/packages/client/src/App.tsx
@@ -4,8 +4,8 @@ import { useMUD } from "./MUDContext";
 export const App = () => {
   const {
     components: { Counter },
-    singletonEntity,
-    worldSend,
+    systemCalls: { increment, optimisticIncrement },
+    network: { singletonEntity },
   } = useMUD();
 
   const counter = useComponentValue(Counter, singletonEntity);
@@ -19,14 +19,19 @@ export const App = () => {
         type="button"
         onClick={async (event) => {
           event.preventDefault();
-
-          const tx = await worldSend("increment", []);
-
-          console.log("increment tx", tx);
-          console.log("increment result", await tx.wait());
+          console.log("new counter value:", await increment());
         }}
       >
         Increment
+      </button>
+      <button
+        type="button"
+        onClick={async (event) => {
+          event.preventDefault();
+          console.log("new counter value:", await optimisticIncrement());
+        }}
+      >
+        Optimistic increment
       </button>
     </>
   );
