+++ templates/react/packages/client/src/MUDContext.tsx
@@ -3,11 +3,12 @@ import { SetupResult } from "./mud/setup";
 
 const MUDContext = createContext<SetupResult | null>(null);
 
-type Props = SetupResult & {
+type Props = {
   children: ReactNode;
+  value: SetupResult;
 };
 
-export const MUDProvider = ({ children, ...value }: Props) => {
+export const MUDProvider = ({ children, value }: Props) => {
   const currentValue = useContext(MUDContext);
   if (currentValue) throw new Error("MUDProvider can only be used once");
   return <MUDContext.Provider value={value}>{children}</MUDContext.Provider>;
