+++ templates/react/packages/client/src/index.tsx
@@ -11,7 +11,7 @@ const root = ReactDOM.createRoot(rootElement);
 // TODO: figure out if we actually want this to be async or if we should render something else in the meantime
 setup().then((result) => {
   root.render(
-    <MUDProvider {...result}>
+    <MUDProvider value={result}>
       <App />
     </MUDProvider>
   );
