+++ templates/react/packages/client/src/mud/createClientComponents.ts
@@ -1,0 +1,12 @@
+import { overridableComponent } from "@latticexyz/recs";
+import { SetupNetworkResult } from "./setupNetwork";
+
+export type ClientComponents = ReturnType<typeof createClientComponents>;
+
+export function createClientComponents({ components }: SetupNetworkResult) {
+  return {
+    ...components,
+    // add your client components or overrides here
+    Counter: overridableComponent(components.Counter),
+  };
+}
